+++ lib/lantern_ui.ex
@@ -69,6 +69,7 @@ defmodule LanternUI do
     toast: LanternUI.Components.Toast,
     theme: LanternUI.Components.Theme,
     timeline: LanternUI.Components.Timeline,
+    resource_list: LanternUI.Components.ResourceList,
     sheet: LanternUI.Components.Sheet,
     navlist: LanternUI.Components.Navlist,
     loading: LanternUI.Components.Loading,